Cut down on the number of warnings by a marginal amount

+++ modcode/Map.cpp
@@ -2,7 +2,7 @@
 
 vector<Entity*> Map::FindEntities(const string& classname) {
 	vector<Entity*> returnValue;
-	vector<Entity*> mapEnts = qtEntTree.NodesIn(x, y, w, h);
+	vector<Entity*> mapEnts = qtEntTree.NodesIn((float)x, (float)y, (float)w, (float)h);
 	for(auto it = mapEnts.begin(); it != mapEnts.end(); ++it) {
 		if(classname == (*it)->classname) {
 			returnValue.push_back((*it));
